Store column numbers as a u32
It's reasonable to assume that a single column won't exceed four
billion, and this is a small speedup (1% instruction count reduction
on the sample file in #293).

+++ src/diff/dijkstra.rs
@@ -231,7 +231,7 @@ mod tests {
         }]
     }
 
-    fn col_helper(line: usize, col: usize) -> Vec<SingleLineSpan> {
+    fn col_helper(line: usize, col: u32) -> Vec<SingleLineSpan> {
         vec![SingleLineSpan {
             line: line.into(),
             start_col: col,
